backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair>=5.0.0",
    "fastapi>=0.135.1",
    "fpdf2>=2.8.7",
    "httpx>=0.28.1",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.2",
    "pandas>=3.0.1",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "simple-salesforce>=1.12.9",
    "sse-starlette>=3.3.2",
    "streamlit>=1.19.0",
    "supabase>=2.28.2",
    "uvicorn[standard]>=0.41.0",
]
